Ticket #446: Added sample application to mix WAV files using conference bridge
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@1681 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip-apps/src/samples/encdec.c
@@ -201,7 +201,7 @@ static pj_status_t enc_dec_test(const char *codec_id,
 	out_frm.buf = (char*)pcmbuf;
 	out_frm.size = sizeof(pcmbuf);
 	
-	if ((pj_rand() % 100) < lost_pct) {
+	if ((pj_rand() % 100) < (int)lost_pct) {
 	    /* Simulate loss */
 	    CHECK( codec->op->recover(codec, sizeof(pcmbuf), &out_frm) );
 	    TRACE_((THIS_FILE, "%d.%03d Packet lost", T));
